requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
decorator==5.1.1
Django==3.2.17
django-crispy-forms==1.13.0
django-encrypted-model-fields==0.6.5
django-recaptcha==3.0.0
django-sslserver==0.22
fuzzywuzzy==0.18.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
kombu==5.1.0
Levenshtein==0.20.9
mod-wsgi==4.9.4
mysqlclient==2.1.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
polling2==0.5.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.20.9
python-memcached==1.59
pytz==2022.7.1
rapidfuzz==2.11.1
redis==4.3.6
requests==2.27.1
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.3
supervisor==4.2.5
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zipp==3.6.0
